+++ src/pages/ProfilePage.js
@@ -3,13 +3,11 @@ import {useCallback, useEffect, useState} from "react";
 import {Host} from "../BackEnd";
 import Header from "../components/header/Header";
 import {Profile} from "../profile/Profile";
-import {emptyUser} from "../components/header/HeaderAndModals";
 
 export function ProfilePage() {
     const [querySearchParams] = useSearchParams();
     const id = querySearchParams.get("id")
     const [profile, setProfile] = useState({id: "", user_name: "", motto: "", avatar: {url: ""}})
-    const [user, setUser] = useState(emptyUser);
 
     const getProfile = useCallback(function () {
         fetch(Host + "/profile?id=" + id, {
@@ -21,33 +19,18 @@ export function ProfilePage() {
         });
     }, [id])
     useEffect(function () {
-        let usr = localStorage.getItem(userStorageKey);
-        if (usr !== null) {
-            let userName = JSON.parse(localStorage.getItem(userStorageKey)).user;
-            let token = JSON.parse(localStorage.getItem(userStorageKey)).token;
-            setUser({name: userName, token: token});
-        }
         getProfile();
 
     }, [getProfile])
 
-    function logOut() {
-        localStorage.removeItem(userStorageKey)
-        setUser(emptyUser)
-    }
 
     return <>
-        <Header user={user.name} token={user.token} hidePostButton={true} handleLogout={logOut}></Header>
+        <Header hidePostButton={true}></Header>
         <div className={"UserProfile"}>
             {profile.id === undefined || profile.id.trim() === ""
                 ? <p>Profile not loaded (yet?)</p>
-                : <Profile id={profile.id} userName={profile.user_name} motto={profile.motto} url={profile.avatar.url}
-                           currentUser={user.name} userToken={user.token} onUpdate={getProfile}/>
+                : <Profile id={profile.id} userName={profile.user_name} motto={profile.motto} url={profile.avatar.url} onUpdate={getProfile}/>
             }
         </div>
     </>;
 }
-
-
-
-const userStorageKey = "user";